refactor(18150): rename workspace module
