+++ hivemq-edge/src/frontend/src/__test-utils__/react-flow/providers.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react'
-import { EdgeFlowProvider } from '@/modules/EdgeVisualisation/hooks/FlowContext.tsx'
+import { EdgeFlowProvider } from '@/modules/Workspace/hooks/FlowContext.tsx'
 import { ReactFlowProvider } from 'reactflow'
-import { EdgeFlowOptions } from '@/modules/EdgeVisualisation/types.ts'
+import { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
 import 'reactflow/dist/style.css'
 
 export const mockReactFlow = (children: ReactNode, defaults?: Partial<EdgeFlowOptions>) => (
