+++ hivemq-edge/src/frontend/src/__test-utils__/react-flow/utils.ts
@@ -1,7 +1,7 @@
 import { WithCSSVar } from '@chakra-ui/react'
 import { Dict } from '@chakra-ui/utils'
 import { XYPosition } from 'reactflow'
-import { IdStubs } from '@/modules/EdgeVisualisation/types.ts'
+import { IdStubs } from '@/modules/Workspace/types.ts'
 import { MOCK_ADAPTER_ID } from '@/__test-utils__/mocks.ts'
 
 export const MOCK_THEME: Partial<WithCSSVar<Dict>> = {
