+++ hivemq-edge/src/frontend/src/modules/Bridges/components/panels/BridgeEditor.tsx
@@ -15,8 +15,8 @@ import { useEdgeToast } from '@/hooks/useEdgeToast/useEdgeToast.tsx'
 import ConfirmationDialog from '@/components/Modal/ConfirmationDialog.tsx'
 import BridgeMainDrawer from '@/modules/Bridges/components/panels/BridgeMainDrawer.tsx'
 import { bridgeInitialState, useBridgeSetup } from '@/modules/Bridges/hooks/useBridgeConfig.tsx'
-import { NodeTypes } from '@/modules/EdgeVisualisation/types.ts'
-import useWorkspaceStore from '@/modules/EdgeVisualisation/hooks/useWorkspaceStore.ts'
+import { NodeTypes } from '@/modules/Workspace/types.ts'
+import useWorkspaceStore from '@/modules/Workspace/hooks/useWorkspaceStore.ts'
 
 interface BridgeEditorProps {
   isNew?: boolean
