+++ hivemq-edge/src/frontend/src/modules/Workspace/components/controls/GroupNodesControl.tsx
