+++ hivemq-edge/src/frontend/src/modules/Workspace/components/controls/SelectionListener.tsx
