+++ hivemq-edge/src/frontend/src/modules/Workspace/components/drawers/NodePropertyDrawer.spec.cy.tsx
@@ -1,10 +1,10 @@
 /// <reference types="cypress" />
 
-import NodePropertyDrawer from '@/modules/EdgeVisualisation/components/drawers/NodePropertyDrawer.tsx'
+import NodePropertyDrawer from '@/modules/Workspace/components/drawers/NodePropertyDrawer.tsx'
 import { MOCK_NODE_ADAPTER } from '@/__test-utils__/react-flow/nodes.ts'
 import { Node } from 'reactflow'
 import { Adapter, Bridge } from '@/api/__generated__'
-import { NodeTypes } from '@/modules/EdgeVisualisation/types.ts'
+import { NodeTypes } from '@/modules/Workspace/types.ts'
 
 const mockNode: Node<Bridge | Adapter> = {
   position: { x: 0, y: 0 },
