+++ hivemq-edge/src/frontend/src/modules/Workspace/components/drawers/WorkspaceOptionsDrawer.tsx
