+++ hivemq-edge/src/frontend/src/modules/Workspace/components/nodes/NodeBridge.spec.cy.tsx
