+++ hivemq-edge/src/frontend/src/modules/Workspace/components/nodes/NodeEdge.tsx
