+++ hivemq-edge/src/frontend/src/modules/Workspace/components/nodes/NodeGroup.tsx
