+++ hivemq-edge/src/frontend/src/modules/Workspace/components/parts/NodeNameCard.tsx
