+++ hivemq-edge/src/frontend/src/modules/Workspace/components/parts/NodeWrapper.tsx
