+++ hivemq-edge/src/frontend/src/modules/Workspace/hooks/FlowContext.spec.tsx
@@ -2,7 +2,7 @@ import { useContext } from 'react'
 import { describe, it, expect } from 'vitest'
 import { render, act } from '@testing-library/react'
 import { EdgeFlowContext, EdgeFlowProvider } from './FlowContext.tsx'
-import { EdgeFlowOptions } from '@/modules/EdgeVisualisation/types.ts'
+import { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
 
 const optionKeys = ['showTopics', 'showStatus', 'showMetrics', 'showGateway', 'showHosts']
