+++ hivemq-edge/src/frontend/src/modules/Workspace/hooks/useEdgeFlowContext.spec.tsx
@@ -3,7 +3,7 @@ import { describe, expect } from 'vitest'
 
 import { useEdgeFlowContext } from './useEdgeFlowContext.tsx'
 import { EdgeFlowProvider } from './FlowContext.tsx'
-import { EdgeFlowOptions } from '@/modules/EdgeVisualisation/types.ts'
+import { EdgeFlowOptions } from '@/modules/Workspace/types.ts'
 
 const wrapper: React.JSXElementConstructor<{ children: React.ReactElement }> = ({ children }) => (
   <EdgeFlowProvider>{children}</EdgeFlowProvider>
