+++ hivemq-edge/src/frontend/src/modules/Workspace/types.ts
