+++ hivemq-edge/src/frontend/src/modules/Workspace/utils/nodes-utils.spec.ts
