+++ hivemq-edge/src/frontend/src/modules/Workspace/utils/status-utils.ts
