+++ hivemq-edge/src/frontend/src/modules/Workspace/utils/topics-utils.spec.ts
@@ -3,7 +3,7 @@ import { GenericObjectType, RJSFSchema } from '@rjsf/utils'
 
 import { Adapter, ProtocolAdapter } from '@/api/__generated__'
 import { mockBridge } from '@/api/hooks/useGetBridges/__handlers__'
-import { TopicFilter } from '@/modules/EdgeVisualisation/types.ts'
+import { TopicFilter } from '@/modules/Workspace/types.ts'
 
 import { MOCK_PROTOCOL_MODBUS, MOCK_ADAPTER_MODBUS } from '@/__test-utils__/adapters/modbus.ts'
 import { MOCK_PROTOCOL_SIMULATION, MOCK_ADAPTER_SIMULATION } from '@/__test-utils__/adapters/simulation.ts'
