+++ hivemq-edge/src/frontend/src/modules/Workspace/utils/topics-utils.ts
